Separate DOM from Application Logic
